StructuralHash: address Schuyler's review comments
